deepagents-pro/pyproject.toml
[project]
name = "deepagent-pro"
version = "0.1.0"
description = "智能数据分析平台 - 基于 LangChain Deep Agents SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.12",
    "langchain-openai>=0.3",
    "langgraph>=0.4",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.1",
    "pydantic-settings>=2.7",
    "python-docx>=1.1",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "aiosqlite>=0.21",
    "matplotlib>=3.9",
    "plotly>=6",
    "kaleido>=0.2",
    "httpx>=0.28",
    "ddgs>=9.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.48",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deepagent_pro"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
